refactor: katex.d.ts -> typings-external.ts
This way the .d.ts will not be deleted when filse are deleted through
rimraf
